# Conflicts:
#	python/ray/train/_checkpoint.py
#	python/ray/train/tests/test_tune.py
#	python/ray/tune/callback.py
#	python/ray/tune/checkpoint_manager.py
#	python/ray/tune/ray_trial_executor.py
#	python/ray/tune/schedulers/pbt.py
#	python/ray/tune/syncer.py
#	python/ray/tune/tests/test_checkpoint_manager.py
#	python/ray/tune/trial.py

+++ python/ray/ml/train/integrations/huggingface/huggingface_trainer.py
@@ -11,16 +11,16 @@ import torch
 import transformers
 import transformers.modeling_utils
 import transformers.trainer
+from ray.util.ml_utils.checkpoint_manager import _TrackedCheckpoint, CheckpointStorage
 from transformers.trainer import WEIGHTS_NAME, TRAINING_ARGS_NAME
 import transformers.training_args
 from torch.utils.data import Dataset as TorchDataset
 
 from ray import train
-from ray import tune
 from ray.util import PublicAPI, get_node_ip_address
 from ray.ml.checkpoint import Checkpoint
 from ray.ml.config import RunConfig, ScalingConfig
-from ray.ml.constants import EVALUATION_DATASET_KEY, TRAIN_DATASET_KEY
+from ray.ml.constants import EVALUATION_DATASET_KEY, TRAIN_DATASET_KEY, PREPROCESSOR_KEY
 from ray.ml.preprocessor import Preprocessor
 from ray.ml.train.integrations.torch import TorchTrainer
 from ray.ml.trainer import GenDataset
@@ -53,39 +53,54 @@ from ray.tune.utils.file_transfer import delete_on_node, sync_dir_between_nodes
 # TODO(ml-team): Make dir syncing checkpoint logic generic.
 
 
-# The checkpoint is turned into a dict with node ip & path
-# in HuggingFaceTrainer.as_trainable
-# TODO(team-ml): Refactor checkpoint management along with Tune.
-class _DataParallelSyncingCheckpointManager(_DataParallelCheckpointManager):
-    """As _DataParallelCheckpointManager, but syncs the dir instead of serializing."""
+class _SyncedTrackedCheckpoint(_TrackedCheckpoint):
+    def commit(self, path: Optional[Path] = None) -> None:
+        if (
+            self.storage_mode == CheckpointStorage.MEMORY
+            or not path
+            or not isinstance(self.dir_or_data, dict)
+        ):
+            return
 
-    def write_checkpoint(self, checkpoint: Dict):
-        # If inside a Tune Trainable, then checkpoint with Tune.
-        with tune.checkpoint_dir(step=self._latest_checkpoint_id) as checkpoint_dir:
-            source_ip = checkpoint[NODE_IP_KEY]
-            source_path = checkpoint[CHECKPOINT_PATH_ON_NODE_KEY]
-            target_ip = get_node_ip_address()
-            if source_ip == target_ip:
-                # Move contents of source_path, but not source_path
-                # itself. shutil.move is already recursive.
-                for path in Path(source_path).iterdir():
-                    shutil.move(str(path.absolute()), checkpoint_dir)
-                shutil.rmtree(source_path, ignore_errors=True)
-            else:
-                sync_dir_between_nodes(
-                    source_ip=source_ip,
-                    source_path=source_path,
-                    target_ip=target_ip,
-                    target_path=checkpoint_dir,
-                    return_futures=False,
-                    max_size_bytes=None,
-                )
-                delete_on_node(node_ip=source_ip, path=source_path)
-            checkpoint_dir = Path(checkpoint_dir)
-            save_preprocessor_to_dir(self.preprocessor, checkpoint_dir)
-            # add tune checkpoint id
-            with open(checkpoint_dir.joinpath(TUNE_CHECKPOINT_ID), "w") as f:
-                f.write(str(self._latest_checkpoint_id))
+        source_ip = self.dir_or_data[NODE_IP_KEY]
+        source_path = self.dir_or_data[CHECKPOINT_PATH_ON_NODE_KEY]
+        target_ip = get_node_ip_address()
+
+        if source_ip == target_ip:
+            # Move contents of source_path, but not source_path
+            # itself. shutil.move is already recursive.
+            for inner in Path(source_path).iterdir():
+                shutil.move(str(path.absolute()), inner)
+            shutil.rmtree(source_path, ignore_errors=True)
+        else:
+            sync_dir_between_nodes(
+                source_ip=source_ip,
+                source_path=source_path,
+                target_ip=target_ip,
+                target_path=str(path),
+                return_futures=False,
+                max_size_bytes=None,
+            )
+            delete_on_node(node_ip=source_ip, path=source_path)
+        save_preprocessor_to_dir(self.dir_or_data.pop(PREPROCESSOR_KEY, None), path)
+        # add tune checkpoint id
+        with open(path.joinpath(TUNE_CHECKPOINT_ID), "w") as f:
+            f.write(str(self.id))
+
+
+class _DataParallelSyncingCheckpointManager(_DataParallelCheckpointManager):
+    def _process_persistent_checkpoint(self, checkpoint: _TrackedCheckpoint):
+        sync_checkpoint = _SyncedTrackedCheckpoint(
+            dir_or_data=checkpoint.dir_or_data,
+            storage_mode=checkpoint.storage_mode,
+            checkpoint_id=checkpoint.id,
+            metrics=checkpoint.metrics,
+            node_ip=checkpoint.node_ip,
+        )
+
+        super(
+            _DataParallelSyncingCheckpointManager, self
+        )._process_persistent_checkpoint(checkpoint=sync_checkpoint)
 
 
 @PublicAPI(stability="alpha")
@@ -108,8 +123,9 @@ class HuggingFaceTrainer(TorchTrainer):
     All the other datasets will not be split.
 
     Please note that if you use a custom ``transformers.Trainer`` subclass,
-    the ``get_train_dataloader`` method will be overriden to disable distributed
-    sampling, as the dataset will already be sharded.
+    the ``get_train_dataloader`` method will be wrapped around to disable
+    sharding by ``transformers.IterableDatasetShard``, as the dataset will
+    already be sharded on the Ray AIR side.
 
     HuggingFace loggers will be automatically disabled, and the ``local_rank``
     argument in ``TrainingArguments`` will be automatically set. Please note
@@ -117,6 +133,8 @@ class HuggingFaceTrainer(TorchTrainer):
     argument in ``TrainingArguments`` manually - otherwise, an exception
     (segfault) may be thrown. Furthermore, 'steps' value for ``save_strategy``,
     ``logging_strategy`` and ``evaluation_strategy`` is not yet supported.
+
+    This Trainer requires ``transformers>=4.19.0`` package.
 
     Example:
         .. code-block:: python
@@ -250,11 +268,11 @@ class HuggingFaceTrainer(TorchTrainer):
 
         # Functionality required for HuggingFaceTrainer only added in this
         # version
-        if LooseVersion(transformers.__version__) < LooseVersion("4.18.0"):
+        if LooseVersion(transformers.__version__) < LooseVersion("4.19.0"):
             raise RuntimeError(
-                "HuggingFaceTrainer requires transformers>=4.18.0, but you "
+                "HuggingFaceTrainer requires transformers>=4.19.0, but you "
                 f"have {transformers.__version__} which is incompatible. "
-                "Update on all nodes with `pip install -U 'transformers>=4.18.0'`."
+                "Update on all nodes with `pip install -U 'transformers>=4.19.0'`."
             )
 
         self._validate_trainer_init_per_worker(
@@ -325,7 +343,7 @@ class HuggingFaceTrainer(TorchTrainer):
     ) -> Checkpoint:
         """Replace the directory checkpoint with a node ip & path dict checkpoint.
 
-        This dict checkpoint will be used used to sync the directory.
+        This dict checkpoint will be used to sync the directory.
         If we were to use a directory checkpoint directly, it would get deepcopied &
         serialized unnecessarily."""
         with checkpoint.as_directory() as checkpoint_path:
